Move `firefox_install` library methods to helper file
